% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regional.R
\name{spec_chr_plot}
\alias{spec_chr_plot}
\title{Regional Plot}
\usage{
spec_chr_plot(
  g,
  m,
  t,
  chr = 8,
  tag_p = 3,
  y_lab = "p",
  name_gene = "PENK",
  y_min = NULL,
  y_max = NULL,
  y_ticks = NULL,
  cut_ctr = 2,
  sig_line1 = 3,
  sig_line2 = -3,
  color_tissue = NULL
)
}
\arguments{
\item{g}{genedata}

\item{m}{map data}

\item{t}{tissue data}

\item{chr}{chromosome}

\item{tag_p}{where tag above p}

\item{y_lab}{label y axis}

\item{name_gene}{gene name select}

\item{y_min}{min number tick}

\item{y_max}{max number tick}

\item{y_ticks}{how many tick}

\item{cut_ctr}{zoom in/out}

\item{sig_line1}{significant top number}

\item{sig_line2}{significan bottom number}

\item{color_tissue}{color choice}
}
\value{
plot
}
\description{
Regional Plot
}
